+++ inst/scripts/examples/leveraged.R
@@ -1,0 +1,29 @@
+add_data_loader(function() {
+    ##### Interest rates #####
+    store_timeseries(var_name = ".estr",
+                     expr = read_timeseries("ESTR.csv", date_col = "DATE") |>
+                         select("date", contains("EU000A2X2A25")),
+                     postprocess = function(d) { names(d) <- c("date", "ESTR"); d }
+    )
+    store_timeseries(var_name = ".eonia",
+                     expr = read_timeseries("EONIA.csv", date_col = "DATE") |>
+                         select("date", contains("EONIA_TO")),
+                     postprocess = function(d) { names(d) <- c("date", "EONIA"); d }
+    )
+    ##### MSCI #####
+    msci(var_name = "wusa-nt",
+         col_trans = net_idx_trans,
+         file = "WUSA_FULL-NT.xls")
+    msci(var_name = "wusa-gr",
+         col_trans = gross_idx_trans,
+         benchmarks = set_names(names(net_idx_trans), names(gross_idx_trans)),
+         file = "WUSA_FULL-GR.xls")
+    msci(var_name = "wusa-nt-eur",
+         col_trans = net_idx_trans_ccy("EUR"),
+         file = "WUSA_FULL-NT-EUR.xls")
+    msci(var_name = "wusa-gr-eur",
+         col_trans = gross_idx_trans_ccy("EUR"),
+         benchmarks = set_names(names(net_idx_trans_ccy("EUR")), names(gross_idx_trans_ccy("EUR"))),
+         file = "WUSA_FULL-GR-EUR.xls")
+})
+build_all_series()